Implemented extension classes for app and services. Moved all the necessary configurations to those classes keeping Program.cs clean.
